Re-implement drag fill
